Update Code LSA
Removed standardization from abundances, because used nowhere; DEA worked with filtered abundances, not standardization. Originally standardization was to be done in DEA analysis itself, however this caused negative values, which messed with the DEA analysis. Normalization is done instead. Additionally, according to chatGPT; standardization is not really done for DEA, meaning I see no need to implement it.